Moves concurrency tests back to scripts since they require prod setup
